# Improve encode_var/decode_varint performance