* using `put_u8` simplifies the `encode_varint` code. this slightly changed behavior, but it should have no effect on prost. It improve performance while reducing unsafe code.
* inline `decode_varint` and remove some unnecessary unsafe.

bench
```
varint/small/encode     time:   [214.82 ns 214.91 ns 215.00 ns]
                        change: [-69.573% -69.350% -69.126%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) low mild
  3 (3.00%) high mild
  8 (8.00%) high severe

varint/small/decode     time:   [231.53 ns 231.84 ns 232.21 ns]
                        change: [-17.817% -16.309% -14.947%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  6 (6.00%) high mild
  8 (8.00%) high severe

varint/small/encoded_len
                        time:   [144.18 ns 144.37 ns 144.61 ns]
                        change: [-3.0551% -2.0177% -1.0496%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  12 (12.00%) high severe

varint/medium/encode    time:   [1.1107 us 1.1183 us 1.1284 us]
                        change: [-18.071% -17.222% -16.488%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe

varint/medium/decode    time:   [414.66 ns 414.86 ns 415.12 ns]
                        change: [-9.6812% -9.1417% -8.4957%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  8 (8.00%) high severe

varint/medium/encoded_len
                        time:   [143.71 ns 143.79 ns 143.89 ns]
                        change: [-3.8550% -2.6239% -1.6465%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

varint/large/encode     time:   [2.2581 us 2.2611 us 2.2657 us]
                        change: [-3.1001% -2.2864% -1.5194%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  4 (4.00%) high mild
  6 (6.00%) high severe

varint/large/decode     time:   [724.83 ns 725.05 ns 725.27 ns]
                        change: [-15.970% -14.724% -13.584%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe

varint/large/encoded_len
                        time:   [143.66 ns 143.69 ns 143.73 ns]
                        change: [-0.4454% -0.2427% -0.0520%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  15 (15.00%) high severe

varint/mixed/encode     time:   [1.2272 us 1.2283 us 1.2301 us]
                        change: [-39.637% -39.005% -38.424%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  7 (7.00%) high severe

varint/mixed/decode     time:   [551.94 ns 552.91 ns 553.89 ns]
                        change: [-14.778% -14.118% -13.479%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

varint/mixed/encoded_len
                        time:   [144.99 ns 145.68 ns 146.49 ns]
                        change: [-6.6187% -5.6221% -4.6395%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe
```